dashboard/requirements.txt
streamlit==1.16.0
kafka-python==2.0.2
pandas==1.5.1